requirements.txt
Flask
langchain
langchain-community
python-dotenv
beautifulsoup4
requests
openai
sentry-sdk